@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: '>=3.3.0 <4.0.0'
+  sdk: '>=3.4.0 <4.0.0'
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -32,16 +32,16 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.6
-  go_router: ^13.2.0
+  cupertino_icons: ^1.0.8
   flutter_hooks: ^0.20.5
-  hooks_riverpod: ^2.4.10
-  json_annotation: ^4.8.1
-  hive_flutter: ^1.1.0
-  smooth_page_indicator: ^1.1.0
   flutter_html: ^3.0.0-beta.2
-  flutter_screenutil: ^5.9.0
-  logger: ^2.0.2+1
+  flutter_screenutil: ^5.9.3
+  go_router: ^14.2.0
+  hive_flutter: ^1.1.0
+  hooks_riverpod: ^2.5.1
+  json_annotation: ^4.9.0
+  logger: ^2.3.0
+  smooth_page_indicator: ^1.1.0
 
 dev_dependencies:
   flutter_test:
@@ -52,8 +52,8 @@ dev_dependencies:
   # activated in the `analysis_options.yaml` file located at the root of your
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
-  flutter_lints: ^3.0.1
-  build_runner: ^2.4.8
+  build_runner: ^2.4.11
+  flutter_lints: ^4.0.0
   hive_generator: ^2.0.1
 
 # For information on the generic Dart part of this file, see the
