@@ -1,10 +1,11 @@
 import 'package:cutaway/database/table/store_info.dart';
 import 'package:cutaway/database/table/topic_info.dart';
-import 'package:cutaway/model/StoreSummary.dart';
-import 'package:cutaway/model/TopicSummary.dart';
+import 'package:cutaway/model/store_summary.dart';
+import 'package:cutaway/model/topic_summary.dart';
 import 'package:cutaway/router/route_utils.dart';
-import 'package:cutaway/tool/hookHelper.dart';
-import 'package:cutaway/tool/images.dart';
+import 'package:cutaway/utils/extension_util.dart';
+import 'package:cutaway/utils/hook_util.dart';
+import 'package:cutaway/utils/image_util.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
@@ -22,12 +23,13 @@ class StorePage extends StatelessWidget {
         title: const Text('珍貴選品'),
         actions: [
           IconButton(
-            onPressed: () {
-              context.pushNamed(AppPage.notification.name);
-            },
+            onPressed: () => context.pushNamed(AppPage.notification.name),
             icon: Stack(
               children: [
-                Icon(Icons.notifications, color: Colors.yellow[100]),
+                Icon(
+                  Icons.notifications,
+                  color: Colors.yellow[100],
+                ),
                 const Icon(Icons.notifications_none),
                 Positioned(
                   top: 4,
@@ -35,9 +37,12 @@ class StorePage extends StatelessWidget {
                   child: Container(
                     height: 8,
                     width: 8,
-                    decoration: const BoxDecoration(shape: BoxShape.circle, color: Colors.red),
+                    decoration: const BoxDecoration(
+                      shape: BoxShape.circle,
+                      color: Colors.red,
+                    ),
                   ),
-                )
+                ),
               ],
             ),
           ),
@@ -55,21 +60,21 @@ class _Body extends StatelessWidget {
     '',
     '',
     [
-      StoreSummary(0, Images.ad1, '', ''),
-      StoreSummary(1, Images.ad2, '', ''),
-      StoreSummary(2, Images.ad3, '', ''),
-      StoreSummary(3, Images.ad4, '', ''),
-      StoreSummary(4, Images.ad5, '', ''),
-      StoreSummary(5, Images.ad6, '', ''),
+      StoreSummary(0, ImageUtil.ad1, '', ''),
+      StoreSummary(1, ImageUtil.ad2, '', ''),
+      StoreSummary(2, ImageUtil.ad3, '', ''),
+      StoreSummary(3, ImageUtil.ad4, '', ''),
+      StoreSummary(4, ImageUtil.ad5, '', ''),
+      StoreSummary(5, ImageUtil.ad6, '', ''),
     ],
   );
 
   final TopicSummary feature = TopicSummary('優選店家', '優選店家', [
-    StoreSummary(0, Images.jellycat, 'JELLYCAT', '來自英國倫敦最富創意的絨毛玩具'),
-    StoreSummary(1, Images.jellycat, 'JELLYCAT', '來自英國倫敦最富創意的絨毛玩具'),
-    StoreSummary(2, Images.jellycat, 'JELLYCAT', '來自英國倫敦最富創意的絨毛玩具'),
-    StoreSummary(3, Images.jellycat, 'JELLYCAT', '來自英國倫敦最富創意的絨毛玩具'),
-    StoreSummary(4, Images.jellycat, 'JELLYCAT', '來自英國倫敦最富創意的絨毛玩具'),
+    StoreSummary(0, ImageUtil.jellycat, 'JELLYCAT', '來自英國倫敦最富創意的絨毛玩具'),
+    StoreSummary(1, ImageUtil.jellycat, 'JELLYCAT', '來自英國倫敦最富創意的絨毛玩具'),
+    StoreSummary(2, ImageUtil.jellycat, 'JELLYCAT', '來自英國倫敦最富創意的絨毛玩具'),
+    StoreSummary(3, ImageUtil.jellycat, 'JELLYCAT', '來自英國倫敦最富創意的絨毛玩具'),
+    StoreSummary(4, ImageUtil.jellycat, 'JELLYCAT', '來自英國倫敦最富創意的絨毛玩具'),
   ]);
 
   final List<TopicInfo> _topics = Hive.box<TopicInfo>(tableTopicInfo).values.toList();
@@ -77,65 +82,64 @@ class _Body extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return ListView.builder(
-        itemCount: _topics.length + 2,
-        itemBuilder: (context, index) {
-          if (index == 0) {
-            return ADPageView(ad);
-          } else if (index == 1) {
-            return FeatureStore(feature);
-          } else {
-            return TopicStore(_topics[index - 2]);
-          }
-        });
+      itemCount: _topics.length + 2,
+      itemBuilder: (context, index) {
+        if (index == 0) {
+          return ADPageView(ad);
+        } else if (index == 1) {
+          return FeatureStore(feature);
+        } else {
+          return TopicStore(_topics[index - 2]);
+        }
+      },
+    );
   }
 }
 
 class ADPageView extends HookWidget {
-  const ADPageView(this.ad, {Key? key}) : super(key: key);
+  const ADPageView(this.ad, {super.key});
 
   final TopicSummary ad;
 
   @override
   Widget build(BuildContext context) {
-    var height = useState(MediaQuery.of(context).size.height);
-    var width = useState(MediaQuery.of(context).size.width);
-    var size = useState(height.value > width.value ? width.value : height.value);
+    final radius = MediaQuery.sizeOf(context).radius;
 
     final controller = usePageController();
     useInterval(() {
-      var currentPage = (controller.page ?? 0).toInt();
+      final currentPage = (controller.page ?? 0).toInt();
 
       controller.animateToPage(
-        currentPage < ad.storeSummarys.length - 1 ? (currentPage + 1) : 0,
+        currentPage < ad.storeSummaries.length - 1 ? (currentPage + 1) : 0,
         duration: const Duration(milliseconds: 350),
         curve: Curves.easeIn,
       );
     }, const Duration(seconds: 5));
 
     return SizedBox(
-      height: size.value * 0.35,
+      height: radius * 0.35,
       child: Stack(
         children: [
           PageView.builder(
-              controller: controller,
-              itemCount: ad.storeSummarys.length,
-              itemBuilder: (context, index) {
-                return _buildAD(context, ad.storeSummarys[index]);
-              }),
+            controller: controller,
+            itemCount: ad.storeSummaries.length,
+            itemBuilder: (context, index) => _buildAD(context, ad.storeSummaries[index]),
+          ),
           Padding(
             padding: const EdgeInsets.all(16),
             child: Align(
               alignment: Alignment.bottomCenter,
               child: SmoothPageIndicator(
-                  controller: controller, // PageController
-                  count: ad.storeSummarys.length,
-                  effect: const WormEffect(
-                    dotHeight: 8,
-                    dotWidth: 8,
-                    spacing: 24,
-                    activeDotColor: Colors.black,
-                  ), // your preferred effect
-                  onDotClicked: (index) {}),
+                controller: controller, // PageController
+                count: ad.storeSummaries.length,
+                effect: const WormEffect(
+                  dotHeight: 8,
+                  dotWidth: 8,
+                  spacing: 24,
+                  activeDotColor: Colors.black,
+                ), // your preferred effect
+                onDotClicked: (index) {},
+              ),
             ),
           ),
         ],
@@ -145,11 +149,14 @@ class ADPageView extends HookWidget {
 
   Widget _buildAD(BuildContext context, StoreSummary summary) {
     return Center(
-        child: Image(
-      image: AssetImage(summary.imgUrl),
-      fit: BoxFit.fill,
-      height: double.infinity,
-    ));
+      child: Image(
+        image: AssetImage(
+          summary.imgUrl,
+        ),
+        fit: BoxFit.fill,
+        height: double.infinity,
+      ),
+    );
   }
 }
 
@@ -160,9 +167,7 @@ class FeatureStore extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    var height = MediaQuery.of(context).size.height;
-    var width = MediaQuery.of(context).size.width;
-    var size = height > width ? width : height;
+    final radius = MediaQuery.sizeOf(context).radius;
 
     return Column(
       mainAxisSize: MainAxisSize.min,
@@ -172,18 +177,22 @@ class FeatureStore extends StatelessWidget {
           padding: const EdgeInsets.all(16.0),
           child: Text(
             feature.title,
-            style: const TextStyle(fontWeight: FontWeight.bold),
+            style: const TextStyle(
+              fontWeight: FontWeight.bold,
+            ),
           ),
         ),
         SizedBox(
-          height: size * 0.4,
+          height: radius * 0.4,
           child: ListView.builder(
             physics: const BouncingScrollPhysics(),
-            padding: const EdgeInsets.only(left: 16.0),
+            padding: const EdgeInsets.only(
+              left: 16.0,
+            ),
             scrollDirection: Axis.horizontal,
             shrinkWrap: true,
-            itemCount: feature.storeSummarys.length,
-            itemBuilder: (BuildContext context, int index) => _buildStore(context, feature.storeSummarys[index]),
+            itemCount: feature.storeSummaries.length,
+            itemBuilder: (BuildContext context, int index) => _buildStore(context, feature.storeSummaries[index]),
           ),
         ),
       ],
@@ -194,11 +203,11 @@ class FeatureStore extends StatelessWidget {
     return SizedBox(
       width: 220,
       child: Padding(
-        padding: const EdgeInsets.only(right: 16, bottom: 8),
+        padding: const EdgeInsets.only(
+          right: 16,
+          bottom: 8,
+        ),
         child: Card(
-          /**
-           * clipBehavior讓圖片不會超出範圍
-           */
           clipBehavior: Clip.antiAliasWithSaveLayer,
           elevation: 4,
           child: Column(
@@ -218,7 +227,10 @@ class FeatureStore extends StatelessWidget {
                       child: Container(
                         height: 30,
                         width: 30,
-                        decoration: const BoxDecoration(shape: BoxShape.circle, color: Colors.white70),
+                        decoration: const BoxDecoration(
+                          shape: BoxShape.circle,
+                          color: Colors.white70,
+                        ),
                         child: InkWell(
                           onTap: () {},
                           child: const Icon(Icons.favorite_border),
@@ -238,17 +250,21 @@ class FeatureStore extends StatelessWidget {
                       maxLines: 1,
                       overflow: TextOverflow.ellipsis,
                       store.title,
-                      style: const TextStyle(fontWeight: FontWeight.bold),
+                      style: const TextStyle(
+                        fontWeight: FontWeight.bold,
+                      ),
                     ),
                     Text(
                       maxLines: 1,
                       overflow: TextOverflow.ellipsis,
                       store.description,
-                      style: TextStyle(fontSize: 12.sp),
+                      style: TextStyle(
+                        fontSize: 12.sp,
+                      ),
                     ),
                   ],
                 ),
-              )
+              ),
             ],
           ),
         ),
@@ -264,9 +280,7 @@ class TopicStore extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    var height = MediaQuery.of(context).size.height;
-    var width = MediaQuery.of(context).size.width;
-    var size = height > width ? width : height;
+    final radius = MediaQuery.sizeOf(context).radius;
 
     return Column(
       mainAxisSize: MainAxisSize.min,
@@ -276,14 +290,18 @@ class TopicStore extends StatelessWidget {
           padding: const EdgeInsets.all(16.0),
           child: Text(
             topics.title,
-            style: const TextStyle(fontWeight: FontWeight.bold),
+            style: const TextStyle(
+              fontWeight: FontWeight.bold,
+            ),
           ),
         ),
         SizedBox(
-          height: size * 0.4,
+          height: radius * 0.4,
           child: ListView.builder(
             physics: const BouncingScrollPhysics(),
-            padding: const EdgeInsets.only(left: 16.0),
+            padding: const EdgeInsets.only(
+              left: 16.0,
+            ),
             scrollDirection: Axis.horizontal,
             shrinkWrap: true,
             itemCount: topics.storeInfoList.length,
@@ -291,16 +309,19 @@ class TopicStore extends StatelessWidget {
           ),
         ),
         Padding(
-          padding: const EdgeInsets.symmetric(horizontal: 16),
+          padding: const EdgeInsets.symmetric(
+            horizontal: 16,
+          ),
           child: TextButton(
             onPressed: () => context,
             style: ButtonStyle(
-              overlayColor: MaterialStateProperty.all(Colors.black12),
-              padding: MaterialStateProperty.all<EdgeInsets>(const EdgeInsets.all(6)),
-              // shape: MaterialStateProperty.all<StadiumBorder>(
-              //     const StadiumBorder(side: BorderSide()))
-              shape: MaterialStateProperty.all(
-                RoundedRectangleBorder(side: const BorderSide(), borderRadius: BorderRadius.circular(4)),
+              overlayColor: WidgetStateProperty.all(Colors.black12),
+              padding: WidgetStateProperty.all<EdgeInsets>(const EdgeInsets.all(6)),
+              shape: WidgetStateProperty.all(
+                RoundedRectangleBorder(
+                  side: const BorderSide(),
+                  borderRadius: BorderRadius.circular(4),
+                ),
               ),
             ),
             child: SizedBox(
@@ -308,12 +329,15 @@ class TopicStore extends StatelessWidget {
               child: Center(
                 child: Text(
                   '更多店家',
-                  style: TextStyle(fontSize: 18.sp, color: Colors.black),
+                  style: TextStyle(
+                    fontSize: 18.sp,
+                    color: Colors.black,
+                  ),
                 ),
               ),
             ),
           ),
-        )
+        ),
       ],
     );
   }
@@ -322,7 +346,10 @@ class TopicStore extends StatelessWidget {
     return SizedBox(
       width: 200,
       child: Padding(
-        padding: const EdgeInsets.only(right: 16, bottom: 8),
+        padding: const EdgeInsets.only(
+          right: 16,
+          bottom: 8,
+        ),
         child: Column(
           crossAxisAlignment: CrossAxisAlignment.start,
           children: [
@@ -343,13 +370,16 @@ class TopicStore extends StatelessWidget {
                     child: Container(
                       height: 30,
                       width: 30,
-                      decoration: const BoxDecoration(shape: BoxShape.circle, color: Colors.white70),
+                      decoration: const BoxDecoration(
+                        shape: BoxShape.circle,
+                        color: Colors.white70,
+                      ),
                       child: InkWell(
                         onTap: () {},
                         child: const Icon(Icons.favorite_border),
                       ),
                     ),
-                  )
+                  ),
                 ],
               ),
             ),
@@ -357,13 +387,17 @@ class TopicStore extends StatelessWidget {
               maxLines: 1,
               overflow: TextOverflow.ellipsis,
               store.title,
-              style: const TextStyle(fontWeight: FontWeight.bold),
+              style: const TextStyle(
+                fontWeight: FontWeight.bold,
+              ),
             ),
             Text(
               maxLines: 1,
               overflow: TextOverflow.ellipsis,
               store.description,
-              style: TextStyle(fontSize: 12.sp),
+              style: TextStyle(
+                fontSize: 12.sp,
+              ),
             ),
           ],
         ),
