@@ -1,5 +1,5 @@
 import 'package:cutaway/router/route_utils.dart';
-import 'package:cutaway/tool/images.dart';
+import 'package:cutaway/utils/image_util.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
@@ -14,9 +14,15 @@ class RegisterPage extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Container(
-      padding: const EdgeInsets.symmetric(vertical: 16, horizontal: 16),
+      padding: const EdgeInsets.symmetric(
+        vertical: 16,
+        horizontal: 16,
+      ),
       decoration: const BoxDecoration(
-        image: DecorationImage(image: AssetImage(Images.background), fit: BoxFit.cover),
+        image: DecorationImage(
+          image: AssetImage(ImageUtil.background),
+          fit: BoxFit.cover,
+        ),
       ),
       child: Scaffold(
         backgroundColor: Colors.transparent,
@@ -45,7 +51,10 @@ class _Body extends StatelessWidget {
         ),
         Text(
           '註冊會員',
-          style: TextStyle(fontSize: 18.sp, fontWeight: FontWeight.bold),
+          style: TextStyle(
+            fontSize: 18.sp,
+            fontWeight: FontWeight.bold,
+          ),
         ),
         const SizedBox(
           height: 32,
@@ -73,25 +82,33 @@ class _Body extends StatelessWidget {
         SizedBox(
           width: double.infinity,
           child: TextButton(
-              onPressed: () => GoRouter.of(context).go(AppPage.home.fullPath),
-              style: ButtonStyle(
-                  padding: MaterialStateProperty.all(const EdgeInsets.all(16)),
-                  backgroundColor: MaterialStateProperty.all(Colors.grey),
-                  shape: MaterialStateProperty.all(RoundedRectangleBorder(borderRadius: BorderRadius.circular(4)))),
-              child: const Text(
-                '下一步',
-                style: TextStyle(color: Colors.white),
+            onPressed: () => context.go(AppPage.home.fullPath),
+            style: ButtonStyle(
+              padding: WidgetStateProperty.all(const EdgeInsets.all(16)),
+              backgroundColor: WidgetStateProperty.all(Colors.grey),
+              shape: WidgetStateProperty.all(RoundedRectangleBorder(
+                borderRadius: BorderRadius.circular(4),
               )),
+            ),
+            child: const Text(
+              '下一步',
+              style: TextStyle(
+                color: Colors.white,
+              ),
+            ),
+          ),
         ),
         const SizedBox(
           height: 16,
         ),
         Center(
           child: TextButton(
-            onPressed: () => GoRouter.of(context).go(AppPage.home.fullPath),
+            onPressed: () => context.go(AppPage.home.fullPath),
             child: const Text(
               '忘記密碼?',
-              style: TextStyle(color: Colors.black),
+              style: TextStyle(
+                color: Colors.black,
+              ),
             ),
           ),
         ),
@@ -140,13 +157,22 @@ class _NameField extends ConsumerWidget {
         ),
         contentPadding: const EdgeInsets.only(left: 12),
         enabledBorder: const OutlineInputBorder(
-          borderSide: BorderSide(color: Colors.black, width: 0.5),
+          borderSide: BorderSide(
+            color: Colors.black,
+            width: 0.5,
+          ),
         ),
         focusedBorder: const OutlineInputBorder(
-          borderSide: BorderSide(color: Colors.black, width: 0.5),
+          borderSide: BorderSide(
+            color: Colors.black,
+            width: 0.5,
+          ),
         ),
         border: const OutlineInputBorder(
-          borderSide: BorderSide(color: Colors.black, width: 0.5),
+          borderSide: BorderSide(
+            color: Colors.black,
+            width: 0.5,
+          ),
         ),
       ),
     );
@@ -167,16 +193,18 @@ class _AppellationPicker extends ConsumerWidget {
       onPressed: () => _showAppellationPicker(context, ref),
       style: ButtonStyle(
         alignment: Alignment.centerLeft,
-        padding: MaterialStateProperty.all(const EdgeInsets.all(12)),
-        side: MaterialStateProperty.all(const BorderSide(color: Colors.black, width: 0.5)),
-        backgroundColor: MaterialStateProperty.all(Colors.white),
-        shape: MaterialStateProperty.all(
-          RoundedRectangleBorder(borderRadius: BorderRadius.circular(4)),
+        padding: WidgetStateProperty.all(const EdgeInsets.all(12)),
+        side: WidgetStateProperty.all(const BorderSide(color: Colors.black, width: 0.5)),
+        backgroundColor: WidgetStateProperty.all(Colors.white),
+        shape: WidgetStateProperty.all(
+          RoundedRectangleBorder(
+            borderRadius: BorderRadius.circular(4),
+          ),
         ),
       ),
       child: Consumer(
         builder: (context, ref, child) {
-          var appellation = ref.watch(_appellation.notifier).state;
+          final appellation = ref.watch(_appellation.notifier).state;
           return Text(
             appellation.isEmpty ? '稱謂' : appellation,
             style: TextStyle(color: appellation.isEmpty ? Colors.grey : Colors.black, fontSize: 16.sp),
@@ -192,18 +220,19 @@ class _AppellationPicker extends ConsumerWidget {
     }
 
     showCupertinoModalPopup(
-        context: context,
-        builder: (context) => Container(
-              height: 190,
-              color: CupertinoColors.white,
-              child: GestureDetector(
-                onTap: () {},
-                child: SafeArea(
-                  top: false,
-                  child: _buildCupertinoPicker(ref),
-                ),
-              ),
-            ));
+      context: context,
+      builder: (context) => Container(
+        height: 190,
+        color: CupertinoColors.white,
+        child: GestureDetector(
+          onTap: () {},
+          child: SafeArea(
+            top: false,
+            child: _buildCupertinoPicker(ref),
+          ),
+        ),
+      ),
+    );
   }
 
   Widget _buildCupertinoPicker(ref) {
@@ -219,9 +248,7 @@ class _AppellationPicker extends ConsumerWidget {
                 ),
               ))
           .toList(),
-      onSelectedItemChanged: (index) {
-        ref.read(_appellation.notifier).state = items[index];
-      },
+      onSelectedItemChanged: (index) => ref.read(_appellation.notifier).state = items[index],
     );
   }
 }
@@ -236,17 +263,26 @@ class _BirthdayPicker extends ConsumerWidget {
       child: TextButton(
         onPressed: () => _showDatePicker(context, ref),
         style: ButtonStyle(
-            alignment: Alignment.centerLeft,
-            padding: MaterialStateProperty.all(const EdgeInsets.all(12)),
-            side: MaterialStateProperty.all(const BorderSide(color: Colors.black, width: 0.5)),
-            backgroundColor: MaterialStateProperty.all(Colors.white),
-            shape: MaterialStateProperty.all(RoundedRectangleBorder(borderRadius: BorderRadius.circular(4)))),
+          alignment: Alignment.centerLeft,
+          padding: WidgetStateProperty.all(const EdgeInsets.all(12)),
+          side: WidgetStateProperty.all(const BorderSide(
+            color: Colors.black,
+            width: 0.5,
+          )),
+          backgroundColor: WidgetStateProperty.all(Colors.white),
+          shape: WidgetStateProperty.all(RoundedRectangleBorder(
+            borderRadius: BorderRadius.circular(4),
+          )),
+        ),
         child: Consumer(
           builder: (context, ref, child) {
-            var birthday = ref.watch(_birthday.notifier).state;
+            final birthday = ref.watch(_birthday.notifier).state;
             return Text(
               birthday.isEmpty ? '生日' : birthday,
-              style: TextStyle(color: birthday.isEmpty ? Colors.grey : Colors.black, fontSize: 16.sp),
+              style: TextStyle(
+                color: birthday.isEmpty ? Colors.grey : Colors.black,
+                fontSize: 16.sp,
+              ),
             );
           },
         ),
@@ -265,13 +301,13 @@ class _BirthdayPicker extends ConsumerWidget {
             SizedBox(
               height: 180,
               child: CupertinoDatePicker(
-                  mode: CupertinoDatePickerMode.date,
-                  initialDateTime: DateTime.now(),
-                  minimumDate: DateTime(1900, 1, 1),
-                  maximumDate: DateTime.now(),
-                  onDateTimeChanged: (value) {
-                    ref.read(_birthday.notifier).state = '${value.year}-${value.month}-${value.day}';
-                  }),
+                mode: CupertinoDatePickerMode.date,
+                initialDateTime: DateTime.now(),
+                minimumDate: DateTime(1900, 1, 1),
+                maximumDate: DateTime.now(),
+                onDateTimeChanged: (value) =>
+                    ref.read(_birthday.notifier).state = '${value.year}-${value.month}-${value.day}',
+              ),
             ),
           ],
         ),
