@@ -1,5 +1,5 @@
 import 'package:cutaway/router/route_utils.dart';
-import 'package:cutaway/tool/images.dart';
+import 'package:cutaway/utils/image_util.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:go_router/go_router.dart';
@@ -12,7 +12,10 @@ class WelcomePage extends StatelessWidget {
     return Container(
       padding: const EdgeInsets.symmetric(vertical: 16, horizontal: 16),
       decoration: const BoxDecoration(
-        image: DecorationImage(image: AssetImage(Images.background), fit: BoxFit.cover),
+        image: DecorationImage(
+          image: AssetImage(ImageUtil.background),
+          fit: BoxFit.cover,
+        ),
       ),
       child: Scaffold(
         backgroundColor: Colors.transparent,
@@ -20,7 +23,7 @@ class WelcomePage extends StatelessWidget {
           children: [
             const Spacer(),
             Image(
-              image: const AssetImage(Images.launch_logo),
+              image: const AssetImage(ImageUtil.launchLogo),
               height: MediaQuery.of(context).size.width / 2,
               width: MediaQuery.of(context).size.width / 2,
             ),
@@ -32,15 +35,11 @@ class WelcomePage extends StatelessWidget {
             SizedBox(
               width: double.infinity,
               child: TextButton(
-                onPressed: () {
-                  context.pushNamed(AppPage.register.name);
-                  // final router = GoRouter.of(context);
-                  // router.go('${router.location}/register');
-                },
+                onPressed: () => context.pushNamed(AppPage.register.name),
                 style: ButtonStyle(
-                    padding: MaterialStateProperty.all(const EdgeInsets.all(16)),
-                    backgroundColor: MaterialStateProperty.all(Colors.yellowAccent),
-                    shape: MaterialStateProperty.all(RoundedRectangleBorder(borderRadius: BorderRadius.circular(4)))),
+                    padding: WidgetStateProperty.all(const EdgeInsets.all(16)),
+                    backgroundColor: WidgetStateProperty.all(Colors.yellowAccent),
+                    shape: WidgetStateProperty.all(RoundedRectangleBorder(borderRadius: BorderRadius.circular(4)))),
                 child: const Text(
                   '立即註冊',
                   style: TextStyle(color: Colors.black),
@@ -51,21 +50,24 @@ class WelcomePage extends StatelessWidget {
               height: 16,
             ),
             TextButton(
-                onPressed: () {
-                  context.pushNamed(AppPage.login.name);
-                  // final router = GoRouter.of(context);
-                  // router.go('${router.location}/login');
-                },
-                child: const Text(
-                  '已有會員？登入',
-                  style: TextStyle(color: Colors.black),
-                )),
+              onPressed: () => context.pushNamed(AppPage.login.name),
+              child: const Text(
+                '已有會員？登入',
+                style: TextStyle(
+                  color: Colors.black,
+                ),
+              ),
+            ),
             const SizedBox(
               height: 16,
             ),
             TextButton(
-                onPressed: () => GoRouter.of(context).go('/home'),
-                child: const Text('以訪客模式登入', style: TextStyle(color: Colors.black, fontWeight: FontWeight.bold))),
+              onPressed: () => context.go('/home'),
+              child: const Text(
+                '以訪客模式登入',
+                style: TextStyle(color: Colors.black, fontWeight: FontWeight.bold),
+              ),
+            ),
             const Spacer(),
           ],
         ),
