@@ -1,7 +1,11 @@
 class Guide {
-  Guide(this.image, this.title, this.description);
-
   String image;
   String title;
   String description;
+
+  Guide(
+    this.image,
+    this.title,
+    this.description,
+  );
 }
