@@ -1,8 +1,30 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 
-class BulletinDialog extends StatelessWidget {
-  const BulletinDialog({super.key});
+void showBulletinDialog(BuildContext context) => showGeneralDialog(
+      context: context,
+      pageBuilder: (context, animation, secondaryAnimation) {
+        /**
+         * 要用Align或Center等Widget包裹才能限制大小
+         */
+        return const Center(
+          child: _BulletinDialog(),
+        );
+      },
+      transitionDuration: const Duration(milliseconds: 500),
+      transitionBuilder: (context, animation, secondaryAnimation, child) {
+        return SlideTransition(
+          position: Tween(
+            begin: const Offset(0, 1),
+            end: const Offset(0, 0),
+          ).animate(animation),
+          child: child,
+        );
+      },
+    );
+
+class _BulletinDialog extends StatelessWidget {
+  const _BulletinDialog({super.key});
 
   final String _content = '1. 因應疫情避免交叉就感染，卡個位目前僅提供線上付款方式。\n\n'
       '2. 若您需無接觸送餐，可寫在備註告知管家\n\n'
@@ -14,59 +36,67 @@ class BulletinDialog extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Material(
-      color: Colors.transparent,
+      shape: RoundedRectangleBorder(
+        borderRadius: BorderRadius.circular(24),
+      ),
       child: SizedBox(
-        height: MediaQuery.of(context).size.height * 0.9,
-        width: MediaQuery.of(context).size.width * 0.9,
+        height: MediaQuery.sizeOf(context).height * 0.9,
+        width: MediaQuery.sizeOf(context).width * 0.9,
         child: Column(
           children: [
             Expanded(
-              child: Container(
-                padding: const EdgeInsets.symmetric(vertical: 8, horizontal: 16),
-                decoration: const BoxDecoration(
-                  color: Colors.white,
-                  borderRadius: BorderRadius.only(
-                    topLeft: Radius.circular(24),
-                    topRight: Radius.circular(24),
+              child: SingleChildScrollView(
+                child: Container(
+                  padding: const EdgeInsets.symmetric(
+                    vertical: 8,
+                    horizontal: 16,
                   ),
-                ),
-                child: Column(
-                  children: [
-                    Stack(
-                      children: [
-                        Positioned.fill(
-                          child: Align(
-                            alignment: Alignment.center,
-                            child: Text(
-                              '卡個位防疫公告',
-                              style: TextStyle(fontWeight: FontWeight.bold, fontSize: 16.sp),
+                  child: Column(
+                    children: [
+                      Stack(
+                        children: [
+                          Positioned.fill(
+                            child: Align(
+                              alignment: Alignment.center,
+                              child: Text(
+                                '卡個位防疫公告',
+                                style: TextStyle(
+                                  fontWeight: FontWeight.bold,
+                                  fontSize: 16.sp,
+                                ),
+                              ),
                             ),
                           ),
-                        ),
-                        Align(
-                          alignment: Alignment.centerRight,
-                          child: IconButton(
-                            // padding: EdgeInsets.zero,
-                            constraints: const BoxConstraints(),
-                            icon: const Icon(Icons.close),
-                            onPressed: () => Navigator.pop(context),
+                          Align(
+                            alignment: Alignment.centerRight,
+                            child: IconButton(
+                              // padding: EdgeInsets.zero,
+                              constraints: const BoxConstraints(),
+                              icon: const Icon(Icons.close),
+                              onPressed: () => Navigator.pop(context),
+                            ),
                           ),
+                        ],
+                      ),
+                      const SizedBox(
+                        height: 12,
+                      ),
+                      Text(
+                        _content,
+                        style: TextStyle(
+                          fontSize: 16.sp,
                         ),
-                      ],
-                    ),
-                    const SizedBox(
-                      height: 12,
-                    ),
-                    Text(
-                      _content,
-                      style: TextStyle(fontSize: 16.sp),
-                    ),
-                  ],
+                      ),
+                    ],
+                  ),
                 ),
               ),
             ),
             Container(
-              padding: const EdgeInsets.symmetric(vertical: 8, horizontal: 16),
+              padding: const EdgeInsets.symmetric(
+                vertical: 8,
+                horizontal: 16,
+              ),
               width: double.infinity,
               decoration: const BoxDecoration(
                 color: Colors.yellow,
