@@ -1,7 +1,7 @@
 part of 'register_page.dart';
 
-var _isNameNotEmpty = StateProvider.autoDispose((ref) => false);
+final _isNameNotEmpty = StateProvider.autoDispose((ref) => false);
 
-var _appellation = StateProvider.autoDispose((ref) => '');
+final _appellation = StateProvider.autoDispose((ref) => '');
 
-var _birthday = StateProvider.autoDispose((ref) => '');
+final _birthday = StateProvider.autoDispose((ref) => '');
