@@ -3,15 +3,13 @@ import 'package:go_router/go_router.dart';
 
 class FadeTransitionPage extends CustomTransitionPage {
   FadeTransitionPage({
-    required LocalKey key,
-    required Widget child,
+    required super.key,
+    required super.child,
   }) : super(
-            key: key,
             transitionsBuilder: (c, animation, a2, child) => FadeTransition(
                   opacity: animation.drive(_curveTween),
                   child: child,
-                ),
-            child: child);
+                ));
 
   static final _curveTween = CurveTween(curve: Curves.easeIn);
 }
